# trading-chats-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "凌期AI辅助期货挑选品种和持仓止盈止损分析程序API文档",
        "title": "凌期AI辅助期货挑选品种和持仓止盈止损分析程序",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/schedules": {
            "get": {
                "description": "获取所有定时任务配置列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "获取所有定时任务配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的定时任务配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "创建定时任务配置",
                "parameters": [
                    {
                        "description": "定时任务配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/schedules/{id}": {
            "delete": {
                "description": "根据ID删除定时任务配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "删除定时任务配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "定时任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/schedules/{id}/logs": {
            "get": {
                "description": "根据定时任务ID获取执行记录列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "获取任务执行日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "定时任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/schedules/{id}/status": {
            "put": {
                "description": "更新定时任务的状态（active/paused）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务"
                ],
                "summary": "更新任务状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "定时任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态 (active/paused)",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/system-config": {
            "get": {
                "description": "获取全局系统配置（包括标题、Logo、动态参数等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system-config"
                ],
                "summary": "获取系统配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SystemConfig"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/system-config/basic": {
            "put": {
                "description": "更新全局系统标题和Logo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system-config"
                ],
                "summary": "保存系统基础配置",
                "parameters": [
                    {
                        "description": "系统基础配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SaveSystemBasicConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/system-config/parameters": {
            "put": {
                "description": "更新全局动态参数配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system-config"
                ],
                "summary": "保存系统动态参数",
                "parameters": [
                    {
                        "description": "系统动态参数配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SaveSystemParametersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "models.SaveSystemBasicConfigRequest": {
            "type": "object",
            "properties": {
                "system_logo": {
                    "type": "string"
                },
                "system_title": {
                    "type": "string"
                }
            }
        },
        "models.SaveSystemParametersRequest": {
            "type": "object",
            "properties": {
                "parameters": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.ScheduleConfig": {
            "type": "object",
            "properties": {
                "created_at": {},
                "cron_expr": {
                    "description": "cron表达式，如 \"*/10 * * * *\"",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "param1": {
                    "description": "参数1",
                    "type": "string"
                },
                "param2": {
                    "description": "参数2",
                    "type": "string"
                },
                "status": {
                    "description": "active: 启用, paused: 暂停",
                    "type": "string"
                },
                "template_id": {
                    "description": "关联的提示词模版ID",
                    "type": "string"
                },
                "updated_at": {}
            }
        },
        "models.SystemConfig": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "system_logo": {
                    "type": "string"
                },
                "system_title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}